Test for recieving transcript from context
